Move Schemas definition for Patpass to common schemas
